[ft] Update help text, add bell schedule page
